requirements.txt
fastapi
uvicorn
jinja2
motor
beanie
python-dotenv
